+++ packages/terra-table/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## Unreleased
 
+* Added
+  * Added `boundingRef` prop for bounded table.
+
+* Changed
+  * Changed flowsheet data grid sections to have sticky title.
+
 ## 5.7.0 - (February 20, 2024)
 
 * Added
@@ -11,7 +17,6 @@
 
 * Changed
   * Changed the default rowMinimumHeight value to auto to fit the row contents.
-  * Changed flowsheet data grid sections to have sticky title.
 
 ## 5.5.0 - (February 1, 2024)